re-enable Unit Test case testShouldSuccessfullyAccessPropertyOfObject
This test case consistently failed in buildbot but was not reproducible when run by hand.
When we finally did repro it by hand, we found it depended on env variable TMPDIR, in a way which was not understood:
* If env variable TMPDIR is unset or if it is set TMPDIR=/tmp (or TMPDIR=/tmp/), then the test case fails.
* If env variable TMPDIR is set to any valid directory path, two (or more) levels down from root /, then the test case passes.
  For example,
  TMPDIR=/private/tmp works.
  TMPDIR=/var/a/bunch/of/guck/generated/by/session/manager/  works.
  TMPDIR=/User/seabuild/home/buildbot/m107-dm2/tmp  works.

When you login to OSX on the console, the system session manager always gives you a TMPDIR, and it is not "/tmp"- instead,
it is a long, pseudo-random, generated path in /var.

Until this date, buildbot on these buildslaves did not have any TMPDIR defined.
Before commiting this change, we also set TMPDIR=$HOME/buildslave/tmp in the buildbot buildslaves /opt/setenv.sh files.

If this re-enabled unit test case now passes in buildbot, then this will confirm that there is some kind of dependency on TMPDIR.